BH1750 Driver completed.
Interface : I2C
Descrpt : Light Sensor
There is no input
Output value is the percentage of the light intensity. 
Before use please change the i2c bus address and may you would like to change the name of the device in the source code.
testapp : It is using the ioctl. So that I created the test application to get value from driver.